devops_iac/local_iac/provisioning_impl.py
import vagrant


def process():
    print('Deploying VM using vagrant')

    vagrantfile = '/vm_config.vagrantfile'

    # Initialize the Vagrant environment
    v = vagrant.Vagrant(vagrantfile)

    # Start the virtual machine
    v.up()

    # Verify that the virtual machine is running
    print(v.status())

    # Stop the virtual machine
    # v.halt()
