requirements.txt
click==8.0.0
python-vagrant==1.0.0